requirements.txt
tweepy==4.10.1
python-dotenv==0.21.0